Added particle effect
Added a particle and particleeffect class to streamline the use of particles.
Use is simple:
Create a new ParticleEffect, set position to where you want particles to originate from. The first parameter is radius for the pixel distance you want the pixels to travel, the second is Color for the color of the particle.

Use the generateParticles() method to generate particle (I'd put it in somewhere like an act method). The parameter is for the amount of particles you want to generate per second.